I updated the xamarin version from 4.4 to 4.7 Because i needs an extender component.
Somehow the demoApp in xero is now on position 0 when we are calling connection to tenant
So i changed the position of our application to grab the tenant[1]
This is in the XEROAPI class method of getting token

Completed the components in the Maps.
Order of the list still showing wrong order though, will do the sorting tomorrow

I turned off the API Calls method because i dont need it, atm.
If you see a warning that;s because i turned it off