A little bit optimization on the API calls when init pins called.
Saved position for the contact for next use, to save API calls.
Internet checking added, but not sure of how useful it will be,
Because google maps doesnt work anyway without internet.
Could be removed in the future.

Fixing bugs with the internet label calls

Things i am unsure about
Does our application update the position in real time?
should we call it everytime to update the position?