Bugs fix and changes
Fixing bugs on UI Maps? seems like it no longer appears,
Applying GetLocationAsync on landing perhaps it will take care of new device bug that fails to read the location,

Adjusting default start date on invoice load to be the past 3 months if the client did not supply it.

Found bug on polyline tried to fix, but not sure if it does take effect or not.
It appears out of nowhere, not sure what the triggers are.

Fixing bug on the last point location, did not adjust based on the setting.